# Conflicts:
#	src/main/java/com/coolwen/experimentplatform/controller/LoginController.java
#	src/main/resources/application.properties

+++ src/main/java/com/coolwen/experimentplatform/controller/ExpModelController.java
@@ -4,6 +4,7 @@ package com.coolwen.experimentplatform.controller;
 import com.coolwen.experimentplatform.model.*;
 import com.coolwen.experimentplatform.model.DTO.KaoHeModelStuDTO;
 import com.coolwen.experimentplatform.service.*;
+import com.sun.org.apache.xpath.internal.operations.Mod;
 import org.apache.shiro.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -48,6 +49,12 @@ public class ExpModelController {
     KaoHeModelScoreService kaoHeModelScoreService;
     @Autowired
     TotalScoreCurrentService totalScoreCurrentService;
+    @Autowired
+    ClazzService clazzService;
+    @Autowired
+    CollegeReportService collegeReportService;
+    @Autowired
+    DockerService dockerService;
 
 
     //查询模块信息页面
@@ -70,6 +77,7 @@ public class ExpModelController {
                       int m_classhour,
                       String m_inurl,
                       MultipartFile m_image,
+                      boolean report_type,
                       HttpServletRequest request,
                       HttpSession session
                       )
@@ -82,6 +90,7 @@ public class ExpModelController {
         expModel.setClasshour(m_classhour);
         expModel.setM_inurl(m_inurl);
         expModel.setImageurl(request.getScheme() + "://" + request.getServerName() + ":" + request.getServerPort()+"/ExperimentPlatform/ExpModelImage/"+file);
+        expModel.setReport_type(report_type);
         expModelService.save(expModel);
         session.setAttribute("modelId",expModel.getM_id());
         return "redirect:/expmodel/addTheory";
@@ -117,8 +126,14 @@ public class ExpModelController {
         moduleTestQuestService.deleteAllModuleTestQuest(moduleTestQuests);
         //删除实验模块报告与实验报告回答
         List<Report> reports = reportService.findReportByMId(id);
-        for(Report r : reports){
-            reportAnswerService.deleteReportAnswerByReportId(r.getReportId());
+        if(reports != null && !reports.isEmpty()){
+            for(Report r : reports){
+                reportAnswerService.deleteReportAnswerByReportId(r.getReportId());
+            }
+        }
+        List<CollegeReport> collegeReportList = collegeReportService.findCollegeReportByMid(id);
+        if(collegeReportList != null && !collegeReportList.isEmpty()){
+            collegeReportService.deleteCollegeList(collegeReportList);
         }
         reportService.deleteReports(reports);
         expModelService.deleteExpModelById(id);
@@ -141,6 +156,7 @@ public class ExpModelController {
                                   int m_classhour,
                                   String m_inurl,
                                   MultipartFile m_image,
+                                  boolean report_type,
                                   HttpServletRequest request,
                                   @PathVariable("id") int id
                                   )
@@ -151,9 +167,19 @@ public class ExpModelController {
         preExpModel.setM_type(m_type);
         preExpModel.setClasshour(m_classhour);
         preExpModel.setM_inurl(m_inurl);
+        preExpModel.setReport_type(report_type);
         String path = fIleService.upload(request,m_image);
         if(path != null){
             preExpModel.setImageurl(request.getScheme() + "://" + request.getServerName() + ":" + request.getServerPort()+"/ExperimentPlatform/ExpModelImage/"+path);
+        }
+        if(report_type == false){
+            //修改为自定义版删学院版答题记录
+            collegeReportService.deleteCollege(id);
+        }else {
+            List<Report> reportList = reportService.findReportByMId(id);
+            for (Report r : reportList){
+                reportAnswerService.deleteReportAnswerByReportId(r.getReportId());
+            }
         }
         expModelService.save(preExpModel);
         return "redirect:/expmodel/list";
@@ -298,16 +324,28 @@ public class ExpModelController {
     @GetMapping("/alltestModel")
     public String alltest(Model model,@RequestParam(value = "pageNum",required = true,defaultValue = "0")int pageNum,HttpSession session){
         Page<ExpModel> page = expModelService.finExpAll(pageNum);
+        Student student = (Student) SecurityUtils.getSubject().getPrincipal();
+        Docker docker = dockerService.findDockerByStu_id(student.getId());
+        if(docker != null){
+            model.addAttribute("docker",docker);
+        }else {
+            model.addAttribute("docker",null);
+        }
         model.addAttribute("list",expModelService.finExpAll(pageNum));
         session.setAttribute("modulePageNum",pageNum);
         session.setAttribute("isAllModule",true);
         return "home_shiyan/all-test";
-//        return "";
     }
     //实验大厅考核模块
     @GetMapping("/kaoheModel")
     public String kaoModelById(Model model,@RequestParam(value = "pageNum",required = true,defaultValue = "0")int pageNum,HttpSession session){
         Student student = (Student) SecurityUtils.getSubject().getPrincipal();
+        Docker docker = dockerService.findDockerByStu_id(student.getId());
+        if(docker != null){
+            model.addAttribute("docker",docker);
+        }else {
+            model.addAttribute("docker",null);
+        }
         Page<KaoHeModelStuDTO> kaohe = kaoheModelService.findKaoheModelStuDto(student.getId(),pageNum);
         model.addAttribute("k",kaohe);
         session.setAttribute("modulePageNum",pageNum);
@@ -329,8 +367,12 @@ public class ExpModelController {
     }
     //模块测试和实验报告汇总页面
     @GetMapping("/moduleList")
-    public String moduleList(Model model,@RequestParam(value = "pageNum",defaultValue = "0",required = true) int pageNum){
+    public String moduleList(Model model,HttpSession session,
+                             @RequestParam(value = "pageNum",defaultValue = "0",required = true) int pageNum){
         model.addAttribute("page1",expModelService.findModelList(pageNum));
+
+        session.removeAttribute("msg2020612");
+
         return "shiyan/lookTestAndReport";
     }
 
@@ -348,5 +390,55 @@ public class ExpModelController {
     }
 
 
+    //首页跳转过来的模块
+    @GetMapping("/home_exp/{id}")
+    public String homeExp(@PathVariable("id") int id,Model model){
+        Student student = (Student) SecurityUtils.getSubject().getPrincipal();
+        if(student.getClassId() != 0) {
+            ClassModel classModel = clazzService.findById(student.getClassId());
+            if (classModel.getClassIscurrent() == false) {
+                //具备考核资格并且为当期
+                if (kaoheModelService.findKaoheModelByMid(id) != null) {
+                    //考核模块
+                    KaoHeModelStuDTO kaoHeModelStuDTO = kaoheModelService.findKaoHeModelStuDTOByStuId(student.getId(), id);
+                    model.addAttribute("k", kaoHeModelStuDTO);
+                    return "home_shiyan/kaohe_copy";
+                }
+            }
+        }
+        //不具备考核资格
+        ExpModel expModel = expModelService.findExpModelByID(id);
+        model.addAttribute("exp",expModel);
+        return "home_shiyan/all-test_copy";
 
+    }
+
+    //继续学习
+    @GetMapping("/contiuneStudy/{id}")
+    public String contiunrStudy(@PathVariable("id")int id){
+        Student student = (Student) SecurityUtils.getSubject().getPrincipal();
+        if(student.getClassId() != 0) {
+            ClassModel classModel = clazzService.findById(student.getClassId());
+            if (classModel.getClassIscurrent() == false) {
+                if(kaoheModelService.findKaoheModelByMid(id) != null){
+                    return "redirect:/expmodel/kaoheModel";
+                }
+            }
+        }
+        return "redirect:/expmodel/alltestModel";
+    }
+
+    //中转站
+    @GetMapping("/homeExpDispatcher/{id}")
+    public String homeExpDispatcher(@PathVariable("id") int id, Model model,HttpSession session){
+
+        Student student = (Student) session.getAttribute("student");
+        //暂时做了修改，如果没有登录，跳转到登录页
+        if(student == null){
+            return "home_page/login";
+        }
+
+        model.addAttribute("disMid",id);
+        return "kuangjia/shiyan";
+    }
 }
